added resources DSL
renamed basic DSL
